# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend for Stitchfolio.",
        "title": "Stitchfolio-backend API docs",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/api/sf/v1",
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/expense-detail/{id}": {
            "get": {
                "description": "Get an expense detail by id",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ExpenseDetail"
                ],
                "summary": "Get ExpenseDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ExpenseDetail id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.ExpenseDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an expense detail",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ExpenseDetail"
                ],
                "summary": "Update ExpenseDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ExpenseDetail id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "expense detail",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.ExpenseDetail"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes an expense detail",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ExpenseDetail"
                ],
                "summary": "Delete ExpenseDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ExpenseDetail id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
        "/expense-tracker/{expenseId}/expense-detail": {
            "get": {
                "description": "Get all expense details for an expense",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ExpenseDetail"
                ],
                "summary": "Get expense details by expense id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expense id",
                        "name": "expenseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responseModel.ExpenseDetail"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.DataResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Saves an expense detail for an expense",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "ExpenseDetail"
                ],
                "summary": "Save ExpenseDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expense id",
                        "name": "expenseId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "expense detail",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestModel.ExpenseDetail"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responseModel.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requestModel.ExpenseDetail": {
            "type": "object",
            "properties": {
                "expenseId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.AuditFields": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.DataResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.ExpenseDetail": {
            "type": "object",
            "properties": {
                "auditFields": {
                    "$ref": "#/definitions/responseModel.AuditFields"
                },
                "expenseId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "price": {
                    "type": "number"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "responseModel.Response": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string"
                },
                "pageMetaData": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}